jak encrypt/decrypt all    works.
introduced the concept of a jakfile for remembering which files you want to encrypt/decrypt also made it point to a passfile. Will add capability of having it simply have the password in there, so that the jakfile knows the files and has the password on a project to project basis. That way the jakfile is the only thing that members need to SECRETLY pass around per project to encrypt/decrypt with ease.

Also:
Added a bunch of failing tests that are TODOes.